方程式の球解処理において、各種設定の数値はdoubleではなくRealで取り扱うように修正した。

+++ Liffom/LiffomTestProject/SolveEquationTest.cs
@@ -84,7 +84,7 @@ namespace GoodSeat.LiffomTestProject
             Formula f = Formula.Parse("3*x^2 - 45*x");
             Variable x = new Variable("x");
             Formula solution = 14.9998d;
-            double error = 0.001d;
+            Numeric error = 0.001d;
             Formula expected = 15d;
             Formula actual;
 
@@ -103,7 +103,7 @@ namespace GoodSeat.LiffomTestProject
             Formula f = Formula.Parse("3*x^2 - 45*x");
             Variable x = new Variable("x");
             Formula solution = 14.9998d;
-            double error = 0.001d;
+            Numeric error = 0.001d;
             Formula expected = 15d;
             Formula actual;
             actual = SolveEquation.GetRoundSolution(f, x, solution, error);
